// src/types/Motors.ts
export type MotorDetailsApi = {
  id: number;
  image: string;
  image2: string;
  image3: string;
  valor: string;
  valorUnidadeReais: number;
  marca: string;
  modelo: string;
  cilidrada: string;
  anoModelo: number;
  combustivel: string;
  codigoFipe: string;
  mesReferencia: string;
  tipoVeiculo: number;
  siglaCombustivel: string;
  dataConsulta: string;
};